moving the files for fall hacks, mm and frosh into their own folders
(cherry picked from commit 934b03fe7d47eab9cd993f5db963e2efe2544595)
