+++ csss-site/src/events/templates/events/fall_hacks/2020/fall_hacks_submissions2020.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block content %}
 {{ block.super }}
-	<link rel="stylesheet" href="{% static 'events_static/fall_hacks_2020/csss_custom.css' %}">
+	<link rel="stylesheet" href="{% static 'events_static/fall_hacks/2020/csss_custom.css' %}">
 	<section id='above-fold'>
 		<header>
 			<h1 class="title is-size-2">Submissions from the Computing Science Student Society Fall Hacks 2020</h1>
