+++ csss-site/src/events/templates/events/mm/2020/mountain_madness2020.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     {% load staticfiles %}
     <link href="https://fonts.googleapis.com/css?family=Bungee" rel="stylesheet">
-    <link rel="stylesheet" href="{% static 'events_static/mm_hackathon2020/mm2020_main.css' %}">
+    <link rel="stylesheet" href="{% static 'events_static/mm/2020/mm2020_main.css' %}">
     <title>Mountain Madness</title>
   </head>
   <body>
@@ -251,10 +251,10 @@
         <h3>You make great events like this happen.</h3>
         <h2><a class='btn' href="sponsorship.pdf">Support Us!</a></h2>
           <a href="https://sfss.org">
-            <img src="{% static 'events_static/mm_hackathon2020/SFSS_COLOR_LOGO.png' %}">
+            <img src="{% static 'events_static/mm/2020/SFSS_COLOR_LOGO.png' %}">
           </a>
           <a href="https://www.sfu.ca/fas.html">
-            <img src="{% static 'events_static/mm_hackathon2020/sfu_logo.png' %}">
+            <img src="{% static 'events_static/mm/2020/sfu_logo.png' %}">
           </a>
         </div>
       </article-->
