+++ csss-site/src/events/templates/events/mm/2021/mountain_madness2021.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     {% load staticfiles %}
     <link href="https://fonts.googleapis.com/css?family=Bungee" rel="stylesheet">
-    <link rel="stylesheet" href="{% static 'events_static/mm_hackathon2021/mm2021_main.css' %}">
+    <link rel="stylesheet" href="{% static 'events_static/mm/2021/mm2021_main.css' %}">
     <link rel="stylesheet" href="{% static 'events_static/custom_css.css' %}">
     <title>Mountain Madness</title>
   </head>
@@ -259,10 +259,10 @@
         <h3>You make great events like this happen.</h3>
         <h2><a class='btn' href="sponsorship.pdf">Support Us!</a></h2>
           <a href="https://sfss.org">
-            <img src="{% static 'events_static/mm_hackathon2021/SFSS_COLOR_LOGO.png' %}">
+            <img src="{% static 'events_static/mm/2021/SFSS_COLOR_LOGO.png' %}">
           </a>
           <a href="https://www.sfu.ca/fas.html">
-            <img src="{% static 'events_static/mm_hackathon2021/sfu_logo.png' %}">
+            <img src="{% static 'events_static/mm/2021/sfu_logo.png' %}">
           </a>
         </div>
       </article-->
