+++ csss-site/src/events/urls/mm.py
@@ -1,0 +1,8 @@
+from django.conf.urls import url
+
+from events.views import mm
+
+urlpatterns = [
+    url(r'^mountain_madness2020$', mm.mountain_madness2020, name='mountain_madness2020'),
+    url(r'^mountain_madness2021$', mm.mountain_madness2021, name='mountain_madness2021'),
+]
