+++ csss-site/src/events/views/frosh/frosh_info.py
@@ -1,0 +1,9 @@
+from django.shortcuts import render
+
+from csss.views_helper import create_main_context
+
+TAB = 'events'
+
+
+def frosh_week(request):
+    return render(request, 'events/frosh/frosh_week.html', create_main_context(request, TAB))
